Move library files into subdirectory and other clean up
