+++ homeassistant/components/levelhome/_lib/protocol.py
@@ -27,5 +27,3 @@ def coerce_is_locked(state: Any) -> bool | None:
     if isinstance(state, bool):
         return state
     return None
-
-
