+++ homeassistant/components/levelhome/api.py
@@ -3,11 +3,11 @@
 from __future__ import annotations
 
 from asyncio import run_coroutine_threadsafe
+
 from aiohttp import ClientSession
 
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import config_entry_oauth2_flow
-
 
 # Library client and errors now live in client.py
