+++ homeassistant/components/levelhome/strings.json
@@ -35,6 +35,9 @@
     },
     "create_entry": {
       "default": "[%key:common::config_flow::create_entry::authenticated%]"
+    },
+    "error": {
+      "invalid_auth": "[%key:common::config_flow::error::invalid_auth%]"
     }
   }
 }
